- Adds deprecation phpdoc for deprecated bootloaders - Starts moving bootloaders to its components
