+++ composer.json
@@ -55,6 +55,7 @@
             "src/DataGrid/src/helpers.php",
             "src/Dumper/src/helpers.php",
             "src/Framework/helpers.php",
+            "src/Framework/polyfill.php",
             "src/Scaffolder/src/helpers.php",
             "src/Stempler/src/helpers.php",
             "src/Translator/src/helpers.php"
