+++ src/AnnotatedRoutes/src/Bootloader/AnnotatedRoutesBootloader.php
@@ -16,8 +16,8 @@ use Spiral\Boot\Bootloader\Bootloader;
 use Spiral\Boot\EnvironmentInterface;
 use Spiral\Boot\MemoryInterface;
 use Spiral\Bootloader\AttributesBootloader;
-use Spiral\Bootloader\ConsoleBootloader;
 use Spiral\Bootloader\Http\RouterBootloader;
+use Spiral\Console\Bootloader\ConsoleBootloader;
 use Spiral\Core\Container\SingletonInterface;
 use Spiral\Router\Command\ResetCommand;
 use Spiral\Router\GroupRegistry;
