+++ src/Framework/polyfill.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+class_alias(
+    \Spiral\Console\Bootloader\ConsoleBootloader::class,
+    \Spiral\Bootloader\ConsoleBootloader::class
+);
+
+class_alias(
+    \Spiral\Tokenizer\Bootloader\TokenizerBootloader::class,
+    \Spiral\Bootloader\TokenizerBootloader::class
+);
